% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_count.R
\name{IDseq_barcode_match}
\alias{IDseq_barcode_match}
\title{Add antibody and sample specific information to the table with barcode
counts.}
\usage{
IDseq_barcode_match(barcode_count_tbl = barcode_count,
  output_filename = "output/data/barcode_count_matched")
}
\description{
The function makes use of the dplyr package 'left_join' option. If there is
a direct match between shared columns of data table and the input tables
antibody_barcode_index and well_barcode_index.
}

