% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_split_data_table.R
\name{IDseq_import_split_data}
\alias{IDseq_import_split_data}
\title{Import splitted FASTQ data into environment}
\usage{

  IDseq_import_split_data(split_data_file_path = "output/data/IDseq_split_data.tsv")
}
\arguments{
\item{.tsv}{file in a directory path.}
}
\description{
Using fread function, the IDseq_split_data.tsv file is imported and columns are named:
"UMI", "Barcode_1", "Barcode_2", "sample_folder"
}
\examples{
IDseq_import_split_data("output/data/IDseq_split_data.tsv")
}

