% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_reads_from_FASTQ.R
\name{IDseq_split_reads}
\alias{IDseq_split_reads}
\title{Split reads from FASTQ files into UMI, antibody barcode and sample barcode sequence.}
\usage{
IDseq_split_reads(fastq_file, matching_distance = 2)
}
\arguments{
\item{fastq_file}{fastq formatted files in a directory path. Default settings from \code{\link{ShortRead::readFastq}}}
}
\value{
Saves a .tsv file with four columns: UMI , Barcode_1 , Barcode_2, sample_ID. Location of file: workingdir/output/data/
}
\description{
First, the IDseq_split_reads function uses shortread package function "readFastq" to load all reads from
a .fastq or .fastq.gz file. 
The fastq files can be zipped (fastq.gz), if doesn`t work have a look at 
if there is a change in the shortread package and file extensions that can be used.  
Then the reads containing an anchor sequence are extracted 
and these reads are split to retrieve - UMI - Barcode_1 - Barcode_2 sequences (and stored in a table with 3 columns)
Finally a column with the sequencing_ID (folder name that contained the fastq file) is added.
}
\examples{
IDseq_split_reads("experiment/data/seq_id/sample-ID.fastq.gz", matching_distance = 2) 
}
\seealso{
for -importing reads from fastq- information ShortRead::readFastq
}

